consumer/requirements.txt
confluent-kafka>=2.3.0
pymongo>=4.6.0
mysql-connector-python>=9.6.0
